~ | Put functions in friends.apy
